chore(remote-control): post-review cleanup
- Collapse 5 sub-50 LOC files into existing siblings:
  RemoteSettings + RemoteBonjourDelegate folded into RemoteControlServer;
  RemoteCommandRequests folded into RemoteCommandHandlers;
  RemoteBrowseDTOs folded into RemoteBrowseHandlers.
  Rename ETag.swift -> RemoteETag.swift and BundleHTTPHandler.swift ->
  RemoteBundleHTTPHandler.swift (type renamed in lockstep) so the
  Remote* prefix is consistent across the directory.
- Replace inline B0xx bug-tracker hashes with semantic comments that
  explain the WHY (queue index remap, NaN volume guard, ETag reset on
  error, search debounce cancellation, entity-track DOM cap).
- Replace try? swallows in body-decoding sites with do/catch +
  Logger.error via a shared decodeBody helper, matching the project's
  error-logging convention.
